check if snapshot files are downloaded first

+++ init/docker-entrypoint.sh
@@ -27,20 +27,28 @@ if [ -z "$PROJECT_ID" ]; then
 fi
 
 # ---------- Download snapshot files from GCP ---------------------------
-echo "Downloading Minimal DB Data Files from GCP"
+if [ ! -f "/data/snapshot_downloaded" ]; then
+    echo "Downloading Minimal DB Data Files from GCP"
 
-gcloud auth activate-service-account --key-file=/serviceaccounts/$SERVICE_ACCOUNT_FILE
-gcloud storage ls gs://mirrornode-db-export/ --billing-project=$PROJECT_ID
+    gcloud auth activate-service-account --key-file=/serviceaccounts/$SERVICE_ACCOUNT_FILE
+    gcloud storage ls gs://mirrornode-db-export/ --billing-project=$PROJECT_ID
 
-mkdir -p /data/download
-export CLOUDSDK_STORAGE_SLICED_OBJECT_DOWNLOAD_MAX_COMPONENTS=1
-export VERSION_NUMBER=$SNAPSHOT_VERSION
+    mkdir -p /data/download
+    export CLOUDSDK_STORAGE_SLICED_OBJECT_DOWNLOAD_MAX_COMPONENTS=1
+    export VERSION_NUMBER=$SNAPSHOT_VERSION
 
-gcloud storage rsync --billing-project=$PROJECT_ID -r -x '.*_part_\d+_\d+_\d+_atma\.csv\.gz$' "gs://mirrornode-db-export/$VERSION_NUMBER/" /data/download
+    gcloud storage rsync --billing-project=$PROJECT_ID -r -x '.*_part_\d+_\d+_\d+_atma\.csv\.gz$' "gs://mirrornode-db-export/$VERSION_NUMBER/" /data/download
 
+    echo "Done downloading files" >> /data/snapshot_downloaded
+
+    echo "Done downloading minimal DB Data Files"
+else
+    echo "Snapshot files already downloaded"
+fi
 # ---------- Download snapshot files from GCP ---------------------------
-
-echo "Done downloading minimal DB Data Files"
 
 # run bootstrap.log
 ./bootstrap.sh $SNAPSHOT_CPU_CORES /data/download > /dev/null 2>> bootstrap.log
+
+echo "Done running bootstrap script"
+exit 1
